refactor: move recaptcha hook to shared hooks folder (temp site)
Made-with: Cursor
